+++ src/server/authentication/recaptcha.html
@@ -1,0 +1,17 @@
+<section id="contact">
+  <div class="content">
+    <div id="form">
+      <form action="" id="contactForm" method="post">
+        <span>Name</span>
+        <input type="text" name="name" class="name" placeholder="Enter your name" tabindex=1 />
+        <span>Email</span>
+        <input type="text" name="email" class="email" placeholder="Enter your email" tabindex=2 />
+        <span id="captcha"></span>
+        <input type="text" name="captcha" class="captcha" maxlength="4" size="4" placeholder="Enter captcha code"
+          tabindex=3 />
+        <span>Message</span>
+        <textarea class="message" placeholder="Enter your message" tabindex=4></textarea>
+        <input type="submit" name="submit" value="Send e-mail" class="submit" tabindex=5>
+      </form>
+    </div>
+</section>
